requirements.txt
ipyparallel>=8.4.1
line_profiler_pycharm>=1.1.0
openpyxl>=3.0.10
pathlib>=1.0.1
zstandard>=0.18.0
numpy>=1.25.2
pandas>=2.0.3
scikit-learn>=1.5.1
scipy>=1.9.3
statsmodels>=0.13.1
nlopt>=2.7.1
plotly>=5.12.0
colorcet>=3.0.1
lxml>=4.9.1
python-pptx>=0.6.22
git+https://github.com/FERMATRICA/fermatrica.git@opensource#egg=fermatrica
git+https://github.com/FERMATRICA/fermatrica_utils.git@opensource#egg=fermatrica_utils
